+++ pkg/mimirtool/analyze/grafana.go
@@ -8,6 +8,7 @@ package analyze
 import (
 	"encoding/json"
 	"fmt"
+	"strings"
 
 	"github.com/grafana/regexp"
 	"github.com/pkg/errors"
@@ -21,12 +22,20 @@ import (
 )
 
 var (
-	lvRegexp                     = regexp.MustCompile(`(?s)label_values\((.+),.+\)`)
-	lvNoQueryRegexp              = regexp.MustCompile(`(?s)label_values\((.+)\)`)
-	qrRegexp                     = regexp.MustCompile(`(?s)query_result\((.+)\)`)
-	validMetricName              = regexp.MustCompile(`^[a-zA-Z_:][a-zA-Z0-9_:]*$`)
-	variableRangeQueryRangeRegex = regexp.MustCompile(`\[\$?\w+?]`)
-	variableSubqueryRangeRegex   = regexp.MustCompile(`\[\$?\w+:\$?\w+?]`)
+	lvRegexp        = regexp.MustCompile(`(?s)label_values\((.+),.+\)`)
+	lvNoQueryRegexp = regexp.MustCompile(`(?s)label_values\((.+)\)`)
+	qrRegexp        = regexp.MustCompile(`(?s)query_result\((.+)\)`)
+	validMetricName = regexp.MustCompile(`^[a-zA-Z_:][a-zA-Z0-9_:]*$`)
+	replacer        = strings.NewReplacer(
+		"$__interval", "5m",
+		"$interval", "5m",
+		"$resolution", "5s",
+		"$__rate_interval", "15s",
+		"$rate_interval", "15s",
+		"$__range", "1d",
+		"${__range_s:glob}", "30",
+		"${__range_s}", "30",
+	)
 )
 
 type MetricsInGrafana struct {
@@ -205,14 +214,8 @@ func metricsFromPanel(panel minisdk.Panel, metrics map[string]struct{}) []error 
 	return parseErrors
 }
 
-func removeVariablesFromRanges(query string) string {
-	query = variableRangeQueryRangeRegex.ReplaceAllLiteralString(query, `[5m]`)
-	query = variableSubqueryRangeRegex.ReplaceAllLiteralString(query, `[5m:1m]`)
-	return query
-}
-
 func parseQuery(query string, metrics map[string]struct{}) error {
-	expr, err := parser.ParseExpr(removeVariablesFromRanges(query))
+	expr, err := parser.ParseExpr(replacer.Replace(query))
 	if err != nil {
 		return err
 	}